+++ new_pysatl_cpd/steps/data_generation_step/data_handlers/data_handler.py
@@ -1,10 +1,10 @@
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 from collections.abc import Iterable
 from typing import Any
 
 from new_pysatl_cpd.steps.step_processor import StepProcessor
 
 
-class DataHandler(ABC, StepProcessor):
+class DataHandler(StepProcessor):
     @abstractmethod
     def get_data(self, **kwargs: Any) -> Iterable[dict[str, float]]: ...
