+++ new_pysatl_cpd/steps/experiment_execution_step/workers/dummy_worker.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Any, Optional
 
 from new_pysatl_cpd.logger import cpd_logger
-from new_pysatl_cpd.steps.test_execution_step.workers.worker import Worker
+from new_pysatl_cpd.steps.experiment_execution_step.workers.worker import Worker
 
 
 class DummyWorker(Worker):
