+++ new_pysatl_cpd/steps/experiment_execution_step/workers/worker.py
@@ -1,10 +1,10 @@
-from abc import ABC, abstractmethod
+from abc import abstractmethod
 from collections.abc import Iterable
 from typing import Any
 
 from new_pysatl_cpd.steps.step_processor import StepProcessor
 
 
-class Worker(ABC, StepProcessor):
+class Worker(StepProcessor):
     @abstractmethod
     def run(self, **kwargs: Any) -> Iterable[dict[str, float]]: ...
